rbd: int -> int64_t on do_export
Prevent 32-bit overflow.

+++ src/rbd.cc
@@ -219,7 +219,7 @@ static int export_read_cb(uint64_t ofs, size_t len, const char *buf, void *arg)
 
 static int do_export(librbd::Image& image, const char *path)
 {
-  int r;
+  int64_t r;
   librbd::image_info_t info;
   bufferlist bl;
   int fd = open(path, O_WRONLY | O_CREAT | O_EXCL, 0644);
